lower blog fetch read timeout to 2 minutes
wget defaults read timeouts to 900 seconds (15 minutes), and retries
to 20. This results in wget, while fetching blogs with bad images, to
get stuck for 5 hours before moving on to the next file on the blog.
On a blog with several bad images, this problem compounds. Lowering
the timeout to 120 seconds lowers the impact to 40 minutes per bad
image.

Note: the timeout clock is reset when wget receives data, so this will
not affect downloading large files that take over 2 minutes to transfer.